+++ grunt/clean.js
@@ -16,6 +16,16 @@ module.exports = function (grunt) {
         cssmin: [
             "wwwroot/css/*",
             "!wwwroot/css/*.min.css"
+        ],
+        "azure-devops-ui": [
+            "js/libs/ui/Core/core.css",
+            "js/libs/ui/Core/override.css",
+            "js/libs/ui/Components/Icon/FabricIcons.css",
+            "js/libs/ui/Components/Icon/FluentIcons.css",
+            "js/libs/ui/Components/Header/Header.css",
+            "js/libs/ui/Components/Button/Button.css",
+            "js/libs/ui/Components/Spinner/Spinner.css",
+            "js/libs/ui/Components/TextField/TextField.css"
         ]
     });