+++ js/querygantt-tab.js
@@ -11,6 +11,7 @@ require({
         "knockout": "./libs/knockout",
         "text": "./libs/text",
         "vis-timeline": "./libs/vis-timeline",
+        "dom-to-image": "./libs/dom-to-image",
         "whatwg-fetch": "./libs/whatwg-fetch",
         "sdk": "./libs/sdk",
         "api": "./libs/api",